Raname timings to benchmarks
